unicase: Update function protoype to match definition
* lib/unicase/special-casing.h (gl_unicase_special_lookup): Gperf 3.1
uses 'size_t' as the datatype for the 'len' parameter in the functions
it generates. Update the prototype specified here to match the newly
generated function.

+++ lib/unicase/special-casing.h
@@ -15,6 +15,7 @@
    You should have received a copy of the GNU Lesser General Public License
    along with this program.  If not, see <http://www.gnu.org/licenses/>.  */
 
+#include <stddef.h>
 
 /* A special casing context.
    A context is negated through x -> -x.  */
@@ -57,4 +58,4 @@ struct special_casing_rule
 };
 
 extern const struct special_casing_rule *
-       gl_unicase_special_lookup (const char *str, unsigned int len);
+       gl_unicase_special_lookup (const char *str, size_t len);
